+++ src/Oro/Bundle/ApiBundle/Tests/Functional/RestJsonApi/InvalidRequestDataTest.php
@@ -4,7 +4,6 @@ namespace Oro\Bundle\ApiBundle\Tests\Functional\RestJsonApi;
 
 use Oro\Bundle\ApiBundle\Tests\Functional\RestJsonApiTestCase;
 use Oro\Bundle\TestFrameworkBundle\Entity\TestProduct;
-use Symfony\Component\HttpFoundation\Response;
 
 class InvalidRequestDataTest extends RestJsonApiTestCase
 {
@@ -20,19 +19,13 @@ class InvalidRequestDataTest extends RestJsonApiTestCase
             [],
             ''
         );
-        self::assertResponseStatusCodeEquals($response, Response::HTTP_BAD_REQUEST);
-        self::assertResponseContentTypeEquals($response, self::JSON_API_CONTENT_TYPE);
-        self::assertEquals(
+        $this->assertResponseContainsValidationError(
             [
-                'errors' => [
-                    [
-                        'status' => Response::HTTP_BAD_REQUEST,
-                        'title'  => 'request data constraint',
-                        'detail' => 'The request data should not be empty'
-                    ]
-                ]
+                'status' => '400',
+                'title'  => 'request data constraint',
+                'detail' => 'The request data should not be empty'
             ],
-            self::jsonToArray($response->getContent())
+            $response
         );
     }
 
@@ -48,15 +41,14 @@ class InvalidRequestDataTest extends RestJsonApiTestCase
             [],
             '{"data": {"type": test"}}'
         );
-        self::assertResponseStatusCodeEquals($response, Response::HTTP_BAD_REQUEST);
-        self::assertResponseContentTypeEquals($response, 'application/json');
-        self::assertEquals(
+        $this->assertResponseContainsValidationError(
             [
-                'code'    => 400,
-                'message' => 'Invalid json message received.'
-                    . ' Parsing error in [1:22]. Expected \'null\'. Got: test'
+                'status' => '400',
+                'title'  => 'bad request http exception',
+                'detail' => 'Invalid json message received.'
+                    . ' Parsing error in [1:22]. Expected \'null\'. Got: test.'
             ],
-            self::jsonToArray($response->getContent())
+            $response
         );
     }
 
@@ -72,15 +64,14 @@ class InvalidRequestDataTest extends RestJsonApiTestCase
             [],
             '{"data": ▿{"type": test"}}'
         );
-        self::assertResponseStatusCodeEquals($response, Response::HTTP_BAD_REQUEST);
-        self::assertResponseContentTypeEquals($response, 'application/json');
-        self::assertEquals(
+        $this->assertResponseContainsValidationError(
             [
-                'code'    => 400,
-                'message' => 'Invalid json message received.'
-                    . ' Parsing error in [1:10]. Unexpected character for value: ?'
+                'status' => '400',
+                'title'  => 'bad request http exception',
+                'detail' => 'Invalid json message received.'
+                    . ' Parsing error in [1:10]. Unexpected character for value: ?.'
             ],
-            self::jsonToArray($response->getContent())
+            $response
         );
     }
 }