+++ src/Oro/Bundle/FilterBundle/Form/Type/Filter/DateRangeFilterType.php
@@ -7,6 +7,10 @@ use Symfony\Component\Form\FormInterface;
 use Symfony\Component\Form\FormView;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
+/**
+ * For filter type for Date range filter.
+ * Time Zone set to UTC, to prevent dates modification and align filter with DB.
+ */
 class DateRangeFilterType extends AbstractDateFilterType
 {
     const NAME = 'oro_type_date_range_filter';
@@ -45,6 +49,7 @@ class DateRangeFilterType extends AbstractDateFilterType
         $resolver->setDefaults(
             [
                 'field_type'       => DateRangeType::class,
+                'time_zone'        => 'UTC',
                 'widget_options'   => [
                     'showDatevariables' => true,
                     'showTime'          => false,