+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Please refer first to [UPGRADE.md](UPGRADE.md) for the most important items that should be addressed before attempting to upgrade or during the upgrade of a vanilla Oro application.
 
 The current file describes significant changes in the code that may affect the upgrade of your customizations.
+## 5.0-alpha.1
+
+### Changed
+
+#### TestFrameworkBundle
+* Public methods `newBrowserTabIsOpened` and `newBrowserTabIsOpenedAndISwitchToIt` are moved from `Oro\Bundle\TestFrameworkBundle\Tests\Behat\Context\OroMainContext` to dedicated context `Oro\Bundle\TestFrameworkBundle\Tests\Behat\Context\BrowserTabContext`.
 
 ## 4.2.2
 
@@ -8,6 +14,9 @@ The current file describes significant changes in the code that may affect the u
 
 #### EntityExtendBundle
 * The `force` option was added to the `oro:entity-extend:update-config` CLI command to avoid accidental execution of it.
+
+#### LocaleBundle
+* The unused service `oro_locale.repository.localization` was removed.
 
 ## 4.2.0 (2020-01-29)
 [Show detailed list of changes](incompatibilities-4-2.md)